refactor: move login modals to entities
