+++ src/entities/user/ui/AuthModal/index.ts
