+++ src/entities/user/ui/ChangePasswordModal/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChangePasswordModal"
