+++ src/entities/user/ui/RegistrationModal/index.ts
