+++ src/entities/user/ui/RegistrationModal/ui/index.ts
