--- src/widgets/ChangePasswordModal/ui/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ChangePasswordModal"
